+++ CHANGES.md
@@ -8,6 +8,8 @@
 
 - `plot_feat` now shows a bar plot for categorical data
 - `bootstrap_stats` added median computation
+- `IdentBody` and `IdentTail` modules, which are placeholders for the body and tail modules in a network, for use when only a head is needed.
+- `NodePredictor` a special `GraphCollapser` which provides a set of predictions per node in a graph, outputting either (batch x predictions x vertices) or (batch x vertices x predictions)
 
 ## Removals